# docs/swagger.yaml
basePath: /api
definitions:
  models.BestSellingProduct:
    properties:
      nama:
        type: string
      qty_terjual:
        type: integer
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.CheckoutItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  models.CheckoutRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CheckoutItem'
        type: array
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  models.SalesReport:
    properties:
      produk_terlaris:
        $ref: '#/definitions/models.BestSellingProduct'
      total_revenue:
        type: integer
      total_transaksi:
        type: integer
    type: object
  models.Transaction:
    properties:
      created_at:
        type: string
      details:
        items:
          $ref: '#/definitions/models.TransactionDetail'
        type: array
      id:
        type: integer
      total_amount:
        type: integer
    type: object
  models.TransactionDetail:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      subtotal:
        type: integer
      transaction_id:
        type: integer
    type: object
host: kasir-app.fadhilaabiyyu.my.id
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Simple Point of Sales API (Bootcamp Golang)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kasir API
  version: "1.0"
paths:
  /categories:
    get:
      description: Get list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: New category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid request body
          schema:
            type: string
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID or Category not found
          schema:
            type: string
        "404":
          description: Invalid ID or Category not found
          schema:
            type: string
      summary: Delete category by ID
      tags:
      - categories
    get:
      description: Get details of a single category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid ID or Category not found
          schema:
            type: string
        "404":
          description: Invalid ID or Category not found
          schema:
            type: string
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid request or Category not found
          schema:
            type: string
        "404":
          description: Invalid request or Category not found
          schema:
            type: string
      summary: Update category by ID
      tags:
      - categories
  /checkout:
    post:
      consumes:
      - application/json
      description: Create a new transaction with multiple items
      parameters:
      - description: Checkout items
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Invalid request body or empty items
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Checkout transaction
      tags:
      - transactions
  /produk:
    get:
      description: Get list of all products. Can filter by name using query parameter.
      parameters:
      - description: Filter products by name (case-insensitive)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: New product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid request body
          schema:
            type: string
      summary: Create a new product
      tags:
      - products
  /produk/{id}:
    delete:
      description: Delete a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID or Product not found
          schema:
            type: string
        "404":
          description: Invalid ID or Product not found
          schema:
            type: string
      summary: Delete product by ID
      tags:
      - products
    get:
      description: Get details of a single product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid ID or Product not found
          schema:
            type: string
        "404":
          description: Invalid ID or Product not found
          schema:
            type: string
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid request or Product not found
          schema:
            type: string
        "404":
          description: Invalid request or Product not found
          schema:
            type: string
      summary: Update product by ID
      tags:
      - products
  /report:
    get:
      description: Get sales summary for a specific date range. If no dates provided,
        returns today's report.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SalesReport'
        "400":
          description: Invalid date format
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get sales report by date range
      tags:
      - reports
  /report/hari-ini:
    get:
      description: Get sales summary for today including total revenue, total transactions,
        and best selling product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SalesReport'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get today's sales report
      tags:
      - reports
schemes:
- https
- http
swagger: "2.0"